Move file, make more refactors
