+++ app/src/main/java/mccode/qdup/RequesterActivity.java
@@ -1,6 +1,5 @@
 package mccode.qdup;
 
-import android.animation.ArgbEvaluator;
 import android.animation.ValueAnimator;
 import android.annotation.SuppressLint;
 import android.app.Activity;
@@ -39,6 +38,7 @@ import mccode.qdup.Utils.Client.ClientListener;
 import mccode.qdup.Utils.Client.ClientWriter;
 import mccode.qdup.Utils.Listeners.MessageListener;
 import mccode.qdup.Utils.Listeners.SearchListener;
+import mccode.qdup.Utils.Listeners.TrackCreatorListener;
 import mccode.qdup.Utils.Messaging.Message;
 import mccode.qdup.Utils.QueueView.RequesterItemTouchHelper;
 import mccode.qdup.Utils.QueueView.RequesterRecyclerListAdapter;
